+++ bazel/src/main/java/com/google/api/codegen/bazel/ApisVisitor.java
@@ -41,6 +41,8 @@ class ApisVisitor extends SimpleFileVisitor<Path> {
   private final Path destDir;
   private final boolean overwrite;
   private final String transport;
+  private final boolean forceTransport;
+  private final String numericEnums;
   private boolean writerMode;
   private final FileWriter fileWriter;
 
@@ -52,6 +54,8 @@ class ApisVisitor extends SimpleFileVisitor<Path> {
       String rawApiTempl,
       boolean overwrite,
       String transport,
+      Boolean forceTransport,
+      String numericEnums,
       FileWriter fileWriter) {
     this.gapicApiTempl = new BazelBuildFileTemplate(gapicApiTempl);
     this.rootApiTempl = new BazelBuildFileTemplate(rootApiTempl);
@@ -59,7 +63,11 @@ class ApisVisitor extends SimpleFileVisitor<Path> {
     this.srcDir = srcDir.normalize();
     this.destDir = destDir.normalize();
     this.overwrite = overwrite;
+    // Indicates a transport was supplied via command line and it should be respected.
+    // If false, any existing transport value pulled by buildozer will take precedent.
+    this.forceTransport = forceTransport;
     this.transport = transport;
+    this.numericEnums = numericEnums;
     this.writerMode = false;
     this.fileWriter =
         (fileWriter != null)
@@ -80,7 +88,7 @@ class ApisVisitor extends SimpleFileVisitor<Path> {
 
     String dirStr = dir.toString();
 
-    ApiVersionedDir bp = new ApiVersionedDir();
+    ApiVersionedDir bp = new ApiVersionedDir(this.forceTransport);
     bazelApiVerPackages.put(dirStr, bp);
     ApiDir bap = new ApiDir();
     bazelApiPackages.put(dirStr, bap);
@@ -184,7 +192,7 @@ class ApisVisitor extends SimpleFileVisitor<Path> {
     System.out.println(
         "Write File [" + tmplType + "]: " + outDir.toString() + File.separator + "BUILD.bazel");
     try {
-      BazelBuildFileView bpv = new BazelBuildFileView(bp, transport);
+      BazelBuildFileView bpv = new BazelBuildFileView(bp, transport, numericEnums);
       fileWriter.write(outFilePath, template.expand(bpv));
     } catch (RuntimeException ex) {
       ex.printStackTrace();