+++ bazel/src/main/java/com/google/api/codegen/bazel/ArgsParser.java
@@ -81,8 +81,17 @@ class ArgsParser {
       overwrite = "false";
     }
     String transport = parsedArgs.get("--transport");
-    if (transport == null) {
+    boolean forceTransport = transport != null && !transport.isBlank();
+    if (!forceTransport) {
+      // If --transport is not provided, default to grpc+rest.
       transport = "grpc+rest";
+    }
+    String numericEnums = parsedArgs.get("--rest_numeric_enums");
+    if (numericEnums == null) {
+      // Note that we set the default to True here, but we only include it in the BUILD file
+      // templates for each language as each language generator's own BUILD rules support the
+      // feature.
+      numericEnums = "True";
     }
 
     Path srcPath = Paths.get(parsedArgs.get("--src")).normalize();
@@ -123,6 +132,8 @@ class ArgsParser {
             : ApisVisitor.readFile(rawApiTempl),
         overwrite.equals("true"),
         transport,
+        forceTransport,
+        numericEnums,
         fileWriter);
   }